add descriptions

+++ mgnify_pipelines_toolkit/analysis/assembly/summarise_sanntis_bgcs.py
@@ -17,11 +17,21 @@
 import argparse
 import fileinput
 import logging
+import os.path
+
 import pandas as pd
 
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(levelname)s: %(message)s"
 )
+
+"""
+Script parses sanntis GFF output and adds descriptions of annotated MIBiGs classes.
+
+Descriptions were provided by Santiago Fragoso in doc with links to papers
+https://docs.google.com/document/d/1IIGokcEpX_yLjg1H-w4_VAbekGRl6LgjSvafym3KMz0/edit?tab=t.0
+Those were copied to TSV file constants/bgc_descriptions/sanntis.txt
+"""
 
 
 def parse_args():
@@ -31,12 +41,19 @@ def parse_args():
     parser.add_argument(
         "-o", "--output", help="Sanntis summary output file.", required=True
     )
+    parser.add_argument(
+        "-d",
+        "--descriptions",
+        help="Sanntis descriptions file.",
+        required=False,
+        default="mgnify_pipelines_toolkit/constants/bgc_descriptions/sanntis.txt",
+    )
     args = parser.parse_args()
-    return args.sanntis_gff, args.output
+    return args.sanntis_gff, args.output, args.descriptions
 
 
 def main():
-    input_gff, output_filename = parse_args()
+    input_gff, output_filename, descriptions = parse_args()
     dict_list = []
     with fileinput.hook_compressed(input_gff, "r") as file_in:
         for line in file_in:
@@ -65,7 +82,27 @@ def main():
             .reset_index(name="Count")
         )
         df_grouped = df_grouped.sort_values(by="Count", ascending=False)
-        df_grouped.to_csv(output_filename, sep="\t", index=False)
+
+        if descriptions and os.path.exists(descriptions):
+            df_desc = pd.read_csv(descriptions, sep="\t", header=0)
+            df_desc = df_desc.set_index("MIBiG_class")
+            df_merged = df_grouped.merge(
+                df_desc, left_on="nearest_MIBiG_class", right_index=True, how="left"
+            )
+            df_merged["Description"] = df_merged.apply(
+                lambda row: row["nearest_MIBiG_class"].replace(
+                    "NRP", df_desc.loc["NRP"]["Description"]
+                )
+                if pd.isna(row["Description"]) and "NRP" in row["nearest_MIBiG_class"]
+                else row["Description"],
+                axis=1,
+            )
+            df_merged = df_merged[
+                ["nearest_MIBiG", "nearest_MIBiG_class", "Description", "Count"]
+            ]
+            df_merged.to_csv(output_filename, sep="\t", index=False)
+        else:
+            df_grouped.to_csv(output_filename, sep="\t", index=False)
 
 
 if __name__ == "__main__":
